Changed to interface. Trims methods.
The view is now scene as an API for the RealmManager to interact with
rather then extended features for a Pane object. All of the
functionality of replacing and updating inner panes have been removed
and will be conducted by the Client class instead. Also, the
RealmManager no longer is tightly coupled with how panes for requesting
input are created and simply hands the required input information to the
View to determine the best way to get the request from the client user.